Added updated read_corpus method
I have added _read_corpus() in form of a comment. 
In this method, we will read data from train_input_data.txt. Each line in this file has a sentence, sentiment label, and tense label corresponding to it.
_read_corpus() will read the text data and labels from this txt file and store it in self.data and self.labels.
self.labels has labels from 0 to 3 for each combination of the sentiment and tense label.